rest

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImp.java
@@ -8,10 +8,12 @@ import ru.kata.spring.boot_security.demo.models.Role;
 import ru.kata.spring.boot_security.demo.models.User;
 import ru.kata.spring.boot_security.demo.repositories.UserRepository;
 import ru.kata.spring.boot_security.demo.repositories.RoleRepository;
+import ru.kata.spring.boot_security.demo.util.UserNotFoundException;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 @Service
@@ -36,12 +38,17 @@ public class UserServiceImp implements UserService {
 
     @Override
     public User getUser(long id) {
-        return userRepository.getById(id);
+        Optional<User> user = userRepository.findById(id);
+        return user.orElseThrow(UserNotFoundException:: new);
     }
 
     @Transactional
     @Override
     public void deleteUser(long id) {
+        Optional<User> user = userRepository.findById(id);
+        if(user.isEmpty()){
+            throw new UserNotFoundException();
+        }
         userRepository.deleteById(id);
     }
